+++ packages/schema/src/idl.rs
@@ -15,7 +15,7 @@ pub const IDL_VERSION: &str = "1.0.0";
 pub struct Api {
     pub contract_name: String,
     pub contract_version: String,
-    pub instantiate: RootSchema,
+    pub instantiate: Option<RootSchema>,
     pub execute: Option<RootSchema>,
     pub query: Option<RootSchema>,
     pub migrate: Option<RootSchema>,
@@ -38,8 +38,10 @@ impl Api {
             responses: self.responses,
         };
 
-        if let Some(metadata) = &mut json_api.instantiate.schema.metadata {
-            metadata.title = Some("InstantiateMsg".to_string());
+        if let Some(instantiate) = &mut json_api.instantiate {
+            if let Some(metadata) = &mut instantiate.schema.metadata {
+                metadata.title = Some("InstantiateMsg".to_string());
+            }
         }
         if let Some(execute) = &mut json_api.execute {
             if let Some(metadata) = &mut execute.schema.metadata {
@@ -72,7 +74,7 @@ pub struct JsonApi {
     contract_name: String,
     contract_version: String,
     idl_version: String,
-    instantiate: RootSchema,
+    instantiate: Option<RootSchema>,
     execute: Option<RootSchema>,
     query: Option<RootSchema>,
     migrate: Option<RootSchema>,
@@ -86,10 +88,14 @@ impl JsonApi {
     }
 
     pub fn to_schema_files(&self) -> Result<Vec<(String, String)>, EncodeError> {
-        let mut result = vec![(
-            "instantiate.json".to_string(),
-            serde_json::to_string_pretty(&self.instantiate)?,
-        )];
+        let mut result = Vec::new();
+
+        if let Some(instantiate) = &self.instantiate {
+            result.push((
+                "instantiate.json".to_string(),
+                serde_json::to_string_pretty(&instantiate)?,
+            ));
+        }
 
         if let Some(execute) = &self.execute {
             result.push((
@@ -140,10 +146,55 @@ pub enum EncodeError {
 
 #[cfg(test)]
 mod tests {
+    use crate::schema_for;
+
     use super::*;
 
     #[test]
     fn version_is_semver() {
         semver::Version::parse(IDL_VERSION).unwrap();
+    }
+
+    #[test]
+    fn to_schema_files_works() {
+        let empty = Api {
+            contract_name: "my_contract".to_string(),
+            contract_version: "1.2.3".to_string(),
+            instantiate: None,
+            execute: None,
+            query: None,
+            migrate: None,
+            sudo: None,
+            responses: None,
+        };
+
+        let files = empty.render().to_schema_files().unwrap();
+        assert_eq!(files, []);
+
+        #[derive(schemars::JsonSchema)]
+        struct TestMsg {}
+
+        let full = Api {
+            contract_name: "my_contract".to_string(),
+            contract_version: "1.2.3".to_string(),
+            instantiate: Some(schema_for!(TestMsg)),
+            execute: Some(schema_for!(TestMsg)),
+            query: Some(schema_for!(TestMsg)),
+            migrate: Some(schema_for!(TestMsg)),
+            sudo: Some(schema_for!(TestMsg)),
+            responses: Some(BTreeMap::from([(
+                "TestMsg".to_string(),
+                schema_for!(TestMsg),
+            )])),
+        };
+
+        let files = full.render().to_schema_files().unwrap();
+        assert_eq!(files.len(), 6);
+        assert_eq!(files[0].0, "instantiate.json");
+        assert_eq!(files[1].0, "execute.json");
+        assert_eq!(files[2].0, "query.json");
+        assert_eq!(files[3].0, "migrate.json");
+        assert_eq!(files[4].0, "sudo.json");
+        assert_eq!(files[5].0, "response_to_TestMsg.json");
     }
 }